Rename compiler/API2 -> compiler/API.
