+++ compiler/src/iree/compiler/API/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/API2/BUILD                                        #
+# compiler/src/iree/compiler/API/BUILD                                        #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -31,12 +31,12 @@ iree_cc_library(
     MLIRCAPILinalg
     MLIRCAPIPDL
     MLIRCAPITransformDialect
-    iree::compiler::API2::Internal::Embed
-    iree::compiler::API2::Internal::IREECompileToolEntryPoint
-    iree::compiler::API2::Internal::IREEMLIRLSPServerToolEntryPoint
-    iree::compiler::API2::Internal::IREEOptToolEntryPoint
-    iree::compiler::API2::Internal::LLDToolEntryPoint
-    iree::compiler::API2::Internal::MLIRInterop
+    iree::compiler::API::Internal::Embed
+    iree::compiler::API::Internal::IREECompileToolEntryPoint
+    iree::compiler::API::Internal::IREEMLIRLSPServerToolEntryPoint
+    iree::compiler::API::Internal::IREEOptToolEntryPoint
+    iree::compiler::API::Internal::LLDToolEntryPoint
+    iree::compiler::API::Internal::MLIRInterop
   PUBLIC
 )
 
@@ -65,7 +65,7 @@ if(NOT IREE_ENABLE_ASAN
     AND NOT IREE_ENABLE_MSAN
     AND NOT IREE_ENABLE_TSAN
     AND NOT IREE_ENABLE_UBSAN)
-  target_link_options(iree_compiler_API2_SharedImpl PRIVATE
+  target_link_options(iree_compiler_API_SharedImpl PRIVATE
     $<$<PLATFORM_ID:Linux>:-Wl,--no-undefined>
     $<$<PLATFORM_ID:Darwin>:-Wl,-undefined,error>
   )
@@ -79,13 +79,13 @@ endif()
 # shared library" seems to be the only place this happens, so we special
 # case vs generalizing.
 target_link_libraries(
-  iree_compiler_API2_SharedImpl
+  iree_compiler_API_SharedImpl
   PRIVATE
-    iree_compiler_API2_StaticImpl
+    iree_compiler_API_StaticImpl
 )
 
 set_target_properties(
-  iree_compiler_API2_SharedImpl
+  iree_compiler_API_SharedImpl
   PROPERTIES
     OUTPUT_NAME "IREECompiler"
     VERSION "0"
@@ -103,7 +103,7 @@ set_target_properties(
 )
 
 install(
-  TARGETS iree_compiler_API2_SharedImpl
+  TARGETS iree_compiler_API_SharedImpl
   COMPONENT Compiler
 )
