+++ compiler/src/iree/compiler/API/Internal/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/API2/Internal/BUILD                               #
+# compiler/src/iree/compiler/API/Internal/BUILD                               #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -102,7 +102,7 @@ iree_cc_library(
   DEPS
     MLIRCAPIIR
     MLIRIR
-    iree::compiler::API2::MLIRInteropHeaders
+    iree::compiler::API::MLIRInteropHeaders
     iree::compiler::ConstEval
     iree::compiler::Dialect::VM::IR
     iree::compiler::Dialect::VM::Target::Bytecode
