+++ compiler/src/iree/compiler/API/Internal/Diagnostics.cpp
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/API2/Internal/Diagnostics.h"
+#include "iree/compiler/API/Internal/Diagnostics.h"
 
 #include "llvm/ADT/TypeSwitch.h"
 #include "llvm/Support/raw_ostream.h"
