+++ compiler/src/iree/compiler/API/MLIRInterop.h
@@ -9,8 +9,8 @@
 // Stability: This API, like the MLIR C-API, offers no ABI stability guarantees,
 // and API stability is best effort.
 
-#ifndef IREE_COMPILER_API2_MLIR_INTEROP_H
-#define IREE_COMPILER_API2_MLIR_INTEROP_H
+#ifndef IREE_COMPILER_API_MLIR_INTEROP_H
+#define IREE_COMPILER_API_MLIR_INTEROP_H
 
 #include "mlir-c/Pass.h"
 #include "mlir-c/Support.h"
@@ -75,4 +75,4 @@ MLIR_CAPI_EXPORTED MlirLogicalResult ireeCompilerTranslateModuletoVMBytecode(
 }
 #endif
 
-#endif  // IREE_COMPILER_API2_MLIR_INTEROP_H
+#endif  // IREE_COMPILER_API_MLIR_INTEROP_H
