+++ compiler/src/iree/compiler/API/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/API2/test/BUILD                                   #
+# compiler/src/iree/compiler/API/test/BUILD                                   #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -18,8 +18,8 @@ iree_cc_test(
   DEPS
     IREELLVMIncludeSetup
     iree::base
-    iree::compiler::API2::Impl
-    iree::compiler::API2::MLIRInteropHeaders
+    iree::compiler::API::Impl
+    iree::compiler::API::MLIRInteropHeaders
     iree::compiler::bindings::c::headers
 )
 
@@ -27,8 +27,8 @@ iree_cc_test(
 
 # Move to bin/ directory and systematically name more appropriately.
 # See: https://github.com/openxla/iree/issues/11297
-if(TARGET iree_compiler_API2_test_api-test-binary)
-  set_target_properties(iree_compiler_API2_test_api-test-binary
+if(TARGET iree_compiler_API_test_api-test-binary)
+  set_target_properties(iree_compiler_API_test_api-test-binary
     PROPERTIES
     OUTPUT_NAME "test-iree-compiler-api-test-binary"
     RUNTIME_OUTPUT_DIRECTORY "${PROJECT_BINARY_DIR}/tools"
