+++ compiler/src/iree/compiler/API/test/api-test-main.c
@@ -15,7 +15,7 @@
 #include <stdio.h>
 
 #include "iree/base/api.h"
-#include "iree/compiler/API2/MLIRInterop.h"
+#include "iree/compiler/API/MLIRInterop.h"
 
 static void bytecode_builder_callback(MlirStringRef str, void* userdata) {
   iree_string_builder_t* builder = (iree_string_builder_t*)userdata;
