+++ compiler/src/iree/compiler/Tools/CMakeLists.txt
@@ -232,7 +232,7 @@ iree_cc_library(
   DEPS
     ::init_iree
     LLVMSupport
-    iree::compiler::API2::Internal::Embed
+    iree::compiler::API::Internal::Embed
     iree::compiler::Pipelines
     iree::compiler::bindings::c::headers
   PUBLIC
